Refactoring some method
